requirements.txt
# PDF Processing
pdfplumber==0.9.0
PyPDF2==3.0.1

# Japanese Text Processing
mecab-python3==1.0.6
sudachipy==0.6.7
sudachidict-core==20230927

# Vector Search
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Text Search
whoosh==2.7.4

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utilities
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
python-multipart==0.0.6
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1